PythonProject/sweets-app/frontend/requirements.txt
streamlit==1.51.0
altair==5.3.0
blinker==1.7.0
cachetools==5.3.3
gitpython==3.1.43
packaging==24.1
pillow==10.4.0
protobuf==4.25.3
pydeck==0.9.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4
watchdog==4.0.1
pandas==2.2.3
numpy==1.26.4
requests==2.32.3
PyJWT
